Initial Commit
This is the first commit, of my attempts, to freeform develop a personal website I can not strive to be content with. Using the VueJS framework, and taking a natural approach, I build a website that is natural using the web technologies. No fancy templating with PHP, no fancy loading, even my components for Vue, while kept in separate files, are just constantsJavaScript constants meant to be treated as super globals.

Some other quircks, uses jQuery's ajax methods to load content to avoid CORS issues, which seem to occur in local development when not using a web server, because the uri includes 'file::'? I do not know. And for code demos, uses iframes. Little else is 'hacky'.