Further Incorporation of VueJS
Strived to further incorporate Vue into the functioning of the website, rather than relying on JS manipulations. This was done by writing components and heavily enforcing the main vue instance in the home page. Existing functions were written as methods of the instance, and later split into more specific methods. Splitting the methods allowed more control of the app such as writing a previous post button that loads the last post saved in an array.